Refactoring: made name and requiresProbabilities static attributes of Metric classes also fixed an important typo in RMSE
This is a backwards compatible change, although there is one TODO that would
make breaking changes in the names if implemented

The reasoning behind this is that the name can be useful in multiple contexts,
e.g. for providing a sortingColumnName to GridSearch or for parsing results
in evaluations. It felt strange having to instantiate the metric objects just
to get access to their name