New feature: unsupervised and rule-based models. Part 1: model base classes
First step towards supporting those use cases. The idea is that such models also
require feature engineering, transformations and so on. They should also
be compatible with evaluators and cross validators. Evaluation interfaces
will be adjusted in a following commit.

The changes are essentially backwards compatible as only abstract classes are adjusted
and some args were given a default value. Only custom implementations of PredictorModel
will truly be affected but in a rather non-intrusive way.