Fixes in vector model and rule based model, fixed bug in classification _predict
Previously _applyPostProcessors was called twice during .predict, this commit fixes it

Interfaces of fit for rule-based models were improved, the type annotations in the base classes
were adjusted.

Predicted variable names are now set by the user in rule-based models.
Since multiple methods in VectorModelClassifier use getModelOutputVars (which cannot be set
for rule-based models), the getters were modified such that getModelOutputVars and getPredictedVars
give the same output for rule based models. The possibility of post processing in rule-based models was
removed, which results in modelOutputVars and predictedVars indeed being the same (since the former
can only differ from the latter if post-processing is happening)