Breaking change: Moved methods for post-processing from VectorModel to VectorRegressionModel
For classification models with discrete labels, it does not seem likely that target transformers
and output transformers add any benefit. Keeping them makes the prediction process more complex and
harder to interpret, as taking the argmax of predicted probabilities would not necessarily coincide
with the result of .predict if post-processing is involved.